Use custom hooks for polling apis
